refactor: removing unassigned nodes in graph is now a separate utils method
Refs: #SFB-8
